+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 ### Updated
-- Connection failure error message
+- Improved connection failure error message
+- It is possible to go async between broker.subscribe and subscription.on('message'). This unlocks the possibility of promise support.
 
 ## 3.2.3
 ### Updated